Fix package loading inside notebooks
Using PlutoDevMacros to create a module with notebooks and having those same notebooks import the module leads to instability when loading the notebooks on startup.  Often the package loading fails and then succeeds after being repeated.  To avoid this problem, there is a new utility package called AppoximationUtils.jl which imports TabularRL.jl as well as other useful packages such as FCANN.  Then the part 2 notebooks can load that package and any other individual notebook with useful code.  The NonTabularRL package now imports the ApproximationUtils package as well as all of the part 2 notebooks.  Also since @reexport is used in NonTabularRL.jl, importing that will include everything in the namespace.

The index.html file now contains notebooks for each part of the book separated as well as the tabular overview notebook at the top