Update actor critic algorithms
Streamline description of policy distributions by adding notation for the state visit counts with different arguments.  Generalize some of the code for the binary value function so it can be used in the actor critic algorithms.  Add parameter studies for the actor critic methods with and without eligibility traces for the short corridor gridworld example.