Add new style documentation to Chapter 10
Update action_value function so it takes the same inputs as are available in the problem setup.  Previously, one had to define a new value function that operated on the state instead of the feature vector.  Keyword arguments are now always passed into the original value function passed to the problem directly.