Add feature selection to tile coding
Tile coding now can use a tuple with more values than are used for tile coding.  a feature_inds argument can be used to identify which elements of the tuple should be used for identifying tile features.  Also a named tuple can be used as a state value with the feature_inds then being a tuple of symbols identifying which elements of the state should be used for tile coding.