Return value estimate with gradient
The gradient update function is now expected to return the value function output associated with the parameters and feature vector.  A separate call to the value function itself is now unnecessary which should save some compute time for the neural network approximations which populate the activations as part of the gradient calculation.